explainshield/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymongo[encryption]==4.6.2
motor==3.3.2
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
xgboost==2.0.3
lightgbm==4.3.0
scikit-learn==1.4.1.post1
shap==0.45.0
pandas==2.2.1
numpy==1.26.4
sentence-transformers==2.5.1
chromadb==0.4.24
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
aif360[all]==0.6.1
fairlearn==0.10.0
reportlab==4.1.0
faker==24.0.0
spacy==3.7.4
scipy==1.12.0
networkx==3.2.1
torch==2.2.1
langchain==0.1.12
langchain-community==0.0.28
river==0.20.1
opacus==1.4.1
